+++ src/lib/djerba/plugins/fusion/constants.py
@@ -15,6 +15,7 @@ MIN_FUSION_READS = 'minimum fusion reads'
 # JSON results keys
 TOTAL_VARIANTS = "Total variants"
 CLINICALLY_RELEVANT_VARIANTS = "Clinically relevant variants"
+NCCN_RELEVANT_VARIANTS = "NCCN relevant variants"
 BODY = 'body'
 FRAME = 'frame'
 GENE = 'gene'
@@ -30,10 +31,13 @@ ENTRCON_NAME = 'entrez_conversion.txt'
 
 # read files from an input directory and gather information on fusions
 
-DATA_FUSIONS_OLD = 'data_fusions.txt'
+DATA_FUSIONS = 'data_fusions.txt'
+DATA_FUSIONS_ONCOKB = 'data_fusions_oncokb.txt'
 DATA_FUSIONS_ANNOTATED = 'data_fusions_oncokb_annotated.txt'
-DATA_FUSIONS_NCCN_ANNOTATED = 'data_fusions_NCCN.txt'
+DATA_FUSIONS_NCCN = 'data_fusions_NCCN.txt'
+DATA_FUSIONS_NCCN_ANNOTATED = 'data_fusions_NCCN_annotated.txt'
 JSON_TO_BE_COMPRESSED = 'fusion_template_to_be_compressed.json'
 FUSION_INDEX = 3
 HUGO_SYMBOL = 'Hugo_Symbol'
-NCCN_RELEVANT_VARIANTS = 'nccn_relevant_variants'
+NCCN_VARIANTS = 'nccn_variants'
+NCCN_ANNOTATION_FILE = 'NCCN_annotations.txt' # Read from djerba/data